requirements.txt
# requirements.txt

# --- Framework Web y Servidor ---
Flask
Flask-Cors
Flask-Limiter
gunicorn

# --- Cliente de OpenAI ---
openai
packaging

# --- Utilidades ---
python-dotenv
httpx
firebase-admin
pyyaml
pydantic
guardrails
agents

# --- Clientes de Google Cloud ---
google-cloud-bigquery
